* info/dir.c (insert_text_into_node): use size_t in argument for a length.
* info/nodes.c (get_node_length): return size_t.

* info/man.c (get_manpage_node, xrefs_of_manpage), info/nodes.c
(build_tags_and_nodes, get_tags_of_indirect_tags_table)
(get_file_character_encoding, get_node_length), info/search.c
(search_forward, search_backward, string_in_line): use size_t for
variables representing length, numbers or indices in arrays.

+++ info/dir.c
@@ -23,8 +23,8 @@
 #include "tilde.h"
 
 static void add_menu_to_node (char *contents, size_t size, NODE *node);
-static void insert_text_into_node (NODE *node, long start,
-    char *text, int textlen);
+static void insert_text_into_node (NODE *node, size_t start,
+    char *text, size_t textlen);
 
 static NODE *dir_node = 0;
 
@@ -203,7 +203,7 @@ add_menu_to_node (char *contents, size_t size, NODE *node)
 }
 
 static void
-insert_text_into_node (NODE *node, long start, char *text, int textlen)
+insert_text_into_node (NODE *node, size_t start, char *text, size_t textlen)
 {
   char *contents;
   long end;
